+++ db/migrate/20210420142343_add_employed_as_supply_teacher_to_early_career_payments_eligibilities.rb
@@ -1,0 +1,5 @@
+class AddEmployedAsSupplyTeacherToEarlyCareerPaymentsEligibilities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :early_career_payments_eligibilities, :employed_as_supply_teacher, :boolean
+  end
+end